% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib2acad-package.R
\docType{package}
\name{bib2acad-package}
\alias{bib2acad-package}
\title{bib2acad: Converting Bibtex Records to .md Files for the Hugo Academic
Theme}
\description{
\code{bib2acad} takes a .bib file and generates for each bib record a .md
file. The name of the .md file is generated from the year and entry field. It
has the format \code{year-month-day_key.md}
}
\section{Required parameters}{
 The only function of this package is
\code{bib2acad()} The bib2acad function needs one parameter.

\strong{bibfile}: A string to the path of the .bib file
("path/to/bibfile")
}

\section{Optional parameters}{
 The \code{bib2acad} function has three optional
parameters

\strong{copybib}: TRUE or FALSE, default = TRUE. Should .bib files
generated?

\strong{abstract}: TRUE or FALSE, default = TRUE. Should abstracts also
added to the .md files?

\strong{overwrite}: TRUE or FALSE, default = FALSE. Should .md files with
the same name overwritten? If .bib files are also created, then this
parameter applies to .bib files as well.
}

\section{Easiest usage}{
 Put your .bib file on the top level of your project
directory and call \code{bib2acad("<your-bibfilename>.bib")}

This generates two folders \code{my-md-folder} and \code{my-bib-folder} where the two
types of generated files will be generated.

To learn more about \code{bib2acad}, see the 'usage' vignette with
\href{../doc/usage.html}{\code{vignette("usage", package = "bib2acad")}}.
}

